Hygiene: CI workflow, screenshot harness, versioned build, AppState split
- .github/workflows/app.yml builds and tests the app on macos-15 with the
  latest stable Xcode, caches the SwiftPM checkouts, and uploads the release
  bundle. The tests need no weights or microphone, so the job is honest.
- app/scripts/snap.sh captures the running app's window (clicking first if
  asked) at full size and 1000 px wide, printing the origin and scale so image
  coordinates map back to screen points. It refuses to click unless
  Transcriber is frontmost. app/scripts/click.swift is the CGEvent helper it
  compiles on demand; its binary is ignored.
- build-app.sh stamps CFBundleShortVersionString from app/VERSION (1.1.0) and
  CFBundleVersion from the commit count, both overridable from the
  environment.
- AppState.swift (757 lines) is now the stored state plus small helpers, with
  the queue consumer in AppState+Jobs, the live session in AppState+Recording,
  import / re-run / delete in AppState+Library and weights in AppState+Models.
  Members those files reach went from private to internal; no behaviour
  changed and the suite is unchanged at 161 + 22.
