+++ .github/workflows/app.yml
@@ -1,0 +1,76 @@
+# Builds the native app and runs its test suite on every push and pull request.
+#
+# The tests need no model weights and no microphone: the Core and Store targets
+# are pure logic, and the Engine tests drive the pipeline through fakes. That
+# is the point of the four-layer split, and this job is what keeps it true.
+name: app
+
+on:
+  push:
+    paths:
+      - "app/**"
+      - ".github/workflows/app.yml"
+  pull_request:
+    paths:
+      - "app/**"
+      - ".github/workflows/app.yml"
+
+concurrency:
+  group: app-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  test:
+    name: swift test (macOS)
+    runs-on: macos-15
+    timeout-minutes: 60
+    defaults:
+      run:
+        working-directory: app
+    steps:
+      - uses: actions/checkout@v4
+
+      # Package.swift is swift-tools-version 6.2, which ships with Xcode 26.
+      - name: Select Xcode
+        uses: maxim-lobanov/setup-xcode@v1
+        with:
+          xcode-version: latest-stable
+
+      - name: Swift version
+        run: swift --version
+
+      # WhisperKit and FluidAudio are large; their checkouts and build products
+      # are worth keeping between runs. Keyed on the resolved graph so a
+      # dependency bump invalidates it.
+      - name: Cache SwiftPM
+        uses: actions/cache@v4
+        with:
+          path: |
+            app/.build/checkouts
+            app/.build/repositories
+            app/.build/artifacts
+          key: spm-${{ runner.os }}-${{ hashFiles('app/Package.resolved') }}
+          restore-keys: spm-${{ runner.os }}-
+
+      - name: Build
+        run: swift build
+
+      - name: Test
+        run: swift test 2>&1 | tee test.log | grep -E "error:|failed|Executed|Test run with" || true
+
+      - name: Check result
+        run: |
+          if grep -qE "with [1-9][0-9]* failures|error:" test.log; then
+            echo "tests failed"; exit 1
+          fi
+          grep -qE "Executed [0-9]+ tests" test.log || { echo "no tests ran"; exit 1; }
+
+      - name: Assemble bundle (release)
+        run: ./build-app.sh release
+
+      - name: Upload bundle
+        uses: actions/upload-artifact@v4
+        with:
+          name: Transcriber.app
+          path: app/Transcriber.app
+          if-no-files-found: error
