+++ .gitignore
@@ -22,3 +22,4 @@ models/
 Transcriber*.zip
 # Xcode per-user state regenerated on open
 .swiftpm/
+app/scripts/click
